src/api/requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.30.*
psycopg2-binary==2.9.*
prometheus-client==0.20.*
